fix: Headers aren't resizing correctly on 2x (#6368)
