+++ app/containers/Header/components/HeaderBackButton.tsx
@@ -8,7 +8,7 @@ const styles = StyleSheet.create({
 		...Platform.select({
 			ios: {
 				minWidth: 34,
-				marginLeft: -19
+				marginLeft: -6
 			},
 			android: {
 				marginHorizontal: 0,
