+++ app/containers/Header/components/HeaderButton/Common.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { StackActions, useNavigation } from '@react-navigation/native';
 import { StyleProp, ViewStyle } from 'react-native';
 
-import I18n from '../../i18n';
-import { isIOS } from '../../lib/methods/helpers';
+import I18n from '../../../../i18n';
+import { isIOS } from '../../../../lib/methods/helpers';
 import Container from './HeaderButtonContainer';
 import Item, { IHeaderButtonItem } from './HeaderButtonItem';
-import { useTheme } from '../../theme';
+import { useTheme } from '../../../../theme';
 
 interface IHeaderButtonCommon extends IHeaderButtonItem {
 	navigation?: any; // TODO: Evaluate proper type
