+++ app/containers/Header/index.tsx
@@ -11,12 +11,11 @@ import { HeaderBackButton } from './components/HeaderBackButton';
 
 interface IHeader extends NativeStackHeaderProps {}
 
-const CustomHeader = ({ options, navigation, route }: IHeader) => {
+const Header = ({ options, navigation, route }: IHeader) => {
 	const { headerLeft, headerTitle, headerRight, title } = options;
 	const [rightButtonsWidth, setRightButtonsWidth] = useState<number | null>(null);
 	const isMasterDetail = useAppSelector(state => state.app.isMasterDetail);
 	const { fontScale } = useWindowDimensions();
-
 	// It helps create an empty view to properly align the header when there is no component on the right.
 	// 32.5 is the value I found that makes it work correctly on both platforms.
 	const size = 32.5 * fontScale;
@@ -70,4 +69,4 @@ const CustomHeader = ({ options, navigation, route }: IHeader) => {
 	);
 };
 
-export default CustomHeader;
+export default Header;
