+++ app/lib/methods/helpers/navigation/index.ts
@@ -4,10 +4,10 @@ import { NativeStackNavigationOptions } from '@react-navigation/native-stack';
 import { themes } from '../../../constants';
 import { TSupportedThemes } from '../../../../theme';
 import sharedStyles from '../../../../views/Styles';
+import Header from '../../../../containers/Header';
 
 export const defaultHeader: NativeStackNavigationOptions = {
-	headerTitleAlign: 'left',
-	headerBackButtonDisplayMode: 'minimal'
+	header: Header
 };
 
 export const drawerStyle = {
