+++ app/views/AdminPanelView/index.tsx
@@ -6,7 +6,7 @@ import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 
 import I18n from '../../i18n';
 import StatusBar from '../../containers/StatusBar';
-import * as HeaderButton from '../../containers/HeaderButton';
+import * as HeaderButton from '../../containers/Header/components/HeaderButton';
 import { getUserSelector } from '../../selectors/login';
 import SafeAreaView from '../../containers/SafeAreaView';
 import { AdminPanelStackParamList } from '../../stacks/types';
