+++ app/views/E2EHowItWorksView.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { StyleSheet } from 'react-native';
 import { RouteProp, useNavigation, useRoute } from '@react-navigation/native';
 
-import * as HeaderButton from '../containers/HeaderButton';
+import * as HeaderButton from '../containers/Header/components/HeaderButton';
 import Markdown from '../containers/markdown';
 import SafeAreaView from '../containers/SafeAreaView';
 import I18n from '../i18n';
