+++ app/views/NewServerView/index.tsx
@@ -7,7 +7,7 @@ import { inviteLinksClear } from '../../actions/inviteLinks';
 import { selectServerRequest, serverFinishAdd, serverRequest } from '../../actions/server';
 import Button from '../../containers/Button';
 import FormContainer, { FormContainerInner } from '../../containers/FormContainer';
-import * as HeaderButton from '../../containers/HeaderButton';
+import * as HeaderButton from '../../containers/Header/components/HeaderButton';
 import { TServerHistoryModel } from '../../definitions';
 import I18n from '../../i18n';
 import { useTheme } from '../../theme';
