+++ app/views/RoomsListView/Header/Header.tsx
@@ -72,9 +72,8 @@ const Header = React.memo(
 		} else {
 			subtitle = server?.replace(/(^\w+:|^)\/\//, '');
 		}
-		// improve copy
 		return (
-			<View style={[styles.container]} accessibilityLabel={`${serverName} ${subtitle}`} accessibilityRole='header' accessible>
+			<View style={styles.container} accessibilityLabel={`${serverName} ${subtitle}`} accessibilityRole='header' accessible>
 				<TouchableOpacity onPress={onPress} testID='rooms-list-header-servers-list-button'>
 					<View style={styles.button}>
 						<Text style={[styles.title, { color: colors.fontTitlesLabels }]} numberOfLines={1}>
