+++ app/views/RoomsListView/index.tsx
@@ -14,7 +14,7 @@ import RoomItem from '../../containers/RoomItem';
 import log, { logEvent, events } from '../../lib/methods/helpers/log';
 import I18n from '../../i18n';
 import { closeSearchHeader, openSearchHeader, roomsRequest } from '../../actions/rooms';
-import * as HeaderButton from '../../containers/HeaderButton';
+import * as HeaderButton from '../../containers/Header/components/HeaderButton';
 import StatusBar from '../../containers/StatusBar';
 import ActivityIndicator from '../../containers/ActivityIndicator';
 import { animateNextTransition } from '../../lib/methods/helpers/layoutAnimation';
@@ -53,7 +53,7 @@ import {
 import { Services } from '../../lib/services';
 import { SupportedVersionsExpired } from '../../containers/SupportedVersions';
 import { ChangePasswordRequired } from '../../containers/ChangePasswordRequired';
-import CustomHeader from '../../containers/CustomHeader';
+import Header from '../../containers/Header';
 
 type TNavigation = CompositeNavigationProp<
 	NativeStackNavigationProp<ChatsStackParamList, 'RoomsListView'>,
@@ -422,10 +422,20 @@ class RoomsListView extends React.Component<IRoomsListViewProps, IRoomsListViewS
 		this.setState({ canCreateRoom }, () => this.setHeader());
 	};
 
+	getBadge = () => {
+		const { notificationPresenceCap, supportedVersionsStatus, theme } = this.props;
+		if (supportedVersionsStatus === 'warn') {
+			return <HeaderButton.BadgeWarn color={colors[theme].buttonBackgroundDangerDefault} />;
+		}
+		if (notificationPresenceCap) {
+			return <HeaderButton.BadgeWarn color={colors[theme].userPresenceDisabled} />;
+		}
+		return null;
+	};
+
 	getHeader = (): any => {
 		const { searching, canCreateRoom } = this.state;
-		const { navigation, isMasterDetail, notificationPresenceCap, issuesWithNotifications, supportedVersionsStatus, theme, user } =
-			this.props;
+		const { navigation, isMasterDetail, issuesWithNotifications, supportedVersionsStatus, theme, user } = this.props;
 		if (searching) {
 			return {
 				headerLeft: () => (
@@ -437,16 +447,6 @@ class RoomsListView extends React.Component<IRoomsListViewProps, IRoomsListViewS
 				headerRight: () => null
 			};
 		}
-
-		const getBadge = () => {
-			if (supportedVersionsStatus === 'warn') {
-				return <HeaderButton.BadgeWarn color={colors[theme].buttonBackgroundDangerDefault} />;
-			}
-			if (notificationPresenceCap) {
-				return <HeaderButton.BadgeWarn color={colors[theme].userPresenceDisabled} />;
-			}
-			return null;
-		};
 
 		const disabled = supportedVersionsStatus === 'expired' || user.requirePasswordChange;
 
@@ -461,7 +461,7 @@ class RoomsListView extends React.Component<IRoomsListViewProps, IRoomsListViewS
 							: // @ts-ignore
 							  () => navigation.toggleDrawer()
 					}
-					badge={() => getBadge()}
+					badge={() => this.getBadge()}
 					disabled={disabled}
 				/>
 			),
@@ -894,7 +894,7 @@ class RoomsListView extends React.Component<IRoomsListViewProps, IRoomsListViewS
 		}
 
 		const options = this.getHeader();
-		return <CustomHeader options={options} navigation={this.props.navigation} route={this.props.route} />;
+		return <Header options={options} navigation={this.props.navigation} route={this.props.route} />;
 	};
 
 	renderItem = ({ item }: { item: IRoomItem }) => {
